my_django_setup/myproject/core/pagination.py
from rest_framework.pagination import PageNumberPagination


class StandardResultsPagination(PageNumberPagination):
    """Standard list pagination: `page`, `page_size` (capped)."""

    page_size = 25
    page_size_query_param = "page_size"
    max_page_size = 100
