my_django_setup/myproject/requirements.txt
Django>=6.0,<7
djangorestframework>=3.15
drf-spectacular>=0.27
django-filter>=24.0
Pillow>=10.0
